added settings in root folder
